requirements.txt
playwright==1.49.1
openai==1.58.1
httpx==0.28.1
python-dotenv==1.0.1
loguru==0.7.3
